chore: rm dev code refactor: create labelColors.ts
