+++ src/labelColors.ts
@@ -1,0 +1,30 @@
+const labelColors: { [key: string]: string } = {
+  color1: "#9dc3ff",
+  color2: "#ff9baf",
+  color3: "#ede478",
+  color4: "#6ed7b9",
+  color5: "#ffa1ef",
+  color6: "#c2a7ff",
+  color7: "#81e1fe",
+  color8: "#ffb79a",
+  
+  color9: "#9dc3ff",
+  color10: "#ff9baf",
+  color11: "#ede478",
+  color12: "#6ed7b9",
+  color13: "#ffa1ef",
+  color14: "#c2a7ff",
+  color15: "#81e1fe",
+  color16: "#ffb79a",
+  
+  color17: "#9dc3ff",
+  color18: "#ff9baf",
+  color19: "#ede478",
+  color20: "#6ed7b9",
+  color21: "#ffa1ef",
+  color22: "#c2a7ff",
+  color23: "#81e1fe",
+  color24: "#ffb79a"
+};
+
+export default labelColors;
